implemented player class
Parses season spreadsheet and creates dictionary of each player entry. Moneypuck doesn't track powerplay and penalty kill goals specifically only 5on4 and 4on5 so any goals that are scored on a 5on3 or 4on3 or 6on4 etc. powerplay or shorthanded are under the 'other' category. Annoying but shouldn't affect much since we can just use the "special goals" stat as an approximation for when we generate the projections.